+++ src/lib.rs
@@ -1,6 +1,7 @@
 #![forbid(unsafe_code)]
 
 pub mod chem_env;
+pub mod display;
 pub mod score;
 pub mod scorer;
 pub mod search;